repackage nodetool individual impls to their own .go file
